requirement.txt
argcomplete==3.2.2
asgiref==3.7.2
awscli==1.32.53
boto3==1.34.53
botocore==1.34.53
certifi==2024.2.2
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
Django==5.0.2
djangorestframework==3.14.0
docutils==0.16
durationpy==0.6
hjson==3.1.0
idna==3.6
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
mysqlclient==2.2.4
placebo==0.9.0
pyasn1==0.5.1
PyMySQL==1.1.0
python-dateutil==2.9.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.2
troposphere==4.6.0
tzdata==2024.1
urllib3==2.0.7
Werkzeug==3.0.1
zappa==0.58.0
